move teams dropdown to own component
