+++ src/app/treasure-hunt/treasure-chest/treasure-chest-menu/teams-dropdown/teams-dropdown.component.html
@@ -1,1 +1,27 @@
-<p>teams-dropdown works!</p>
+<div *ngIf="inReport" class="nav-item dropdown">
+    <a class="nav-link dropdown-toggle" (click)="toggleTeams()">
+        Teams
+    </a>
+    <div class="dropdown-menu p-0 teams-dropdown" [ngClass]="{'show': displayTeamDropdown == true}">
+        <div class="list-group">
+            <a class="list-group-item list-group-item-action" *ngFor="let team of teams"
+                [ngClass]="{'active': team.selected}" (click)="setSelectedTeam(team)">{{team.display}}
+                <span class="pull-right pl-2 pr-2" [ngClass]="{'badge badge-pill': team.numCalcs > 0}">
+                    {{team.numCalcs}}
+                </span>
+            </a>
+        </div>
+    </div>
+</div>
+
+<div *ngIf="!inReport">
+    <h5>Select Teams</h5>
+    <div class="list-group list-group-flush">
+        <a class="list-group-item list-group-item-action" *ngFor="let team of teams"
+            [ngClass]="{'active': team.selected}" (click)="setSelectedTeam(team)">{{team.display}}
+            <span class="pull-right pl-2 pr-2" [ngClass]="{'badge badge-pill': team.numCalcs > 0}">
+                {{team.numCalcs}}
+            </span>
+        </a>
+    </div>
+</div>
