+++ src/app/treasure-hunt/treasure-chest/treasure-chest-menu/treasure-chest-menu.component.html
@@ -57,21 +57,7 @@
   </div>
 
   <!--teams. use dropdown in report-->
-  <div *ngIf="inReport" class="nav-item dropdown">
-    <a class="nav-link dropdown-toggle" (click)="toggleTeams()">
-      Teams
-    </a>
-    <div class="dropdown-menu p-0 teams-dropdown" [ngClass]="{'show': displayTeamDropdown == true}">
-      <div class="list-group">
-        <a class="list-group-item list-group-item-action" *ngFor="let team of teams"
-          [ngClass]="{'active': team.selected}" (click)="setSelectedTeam(team)">{{team.display}}
-          <span class="pull-right pl-2 pr-2" [ngClass]="{'badge badge-pill': team.numCalcs > 0}">
-            {{team.numCalcs}}
-          </span>
-        </a>
-      </div>
-    </div>
-  </div>
+  <app-teams-dropdown *ngIf="inReport" [inReport]="inReport" [settings]="settings"></app-teams-dropdown>
 
   <!--equipment. use dropdown in report-->
   <div *ngIf="inReport" class="nav-item dropdown">
@@ -124,15 +110,7 @@
   <div class="card-body p-2 scroll-item">
     <div class="row no-gutters" *ngIf="displayAdditionalFiltersDropdown == 'show'">
       <div class="col-6 pr-1">
-        <h5>Select Teams</h5>
-        <div class="list-group list-group-flush">
-          <a class="list-group-item list-group-item-action" *ngFor="let team of teams"
-            [ngClass]="{'active': team.selected}" (click)="setSelectedTeam(team)">{{team.display}}
-            <span class="pull-right pl-2 pr-2" [ngClass]="{'badge badge-pill': team.numCalcs > 0}">
-              {{team.numCalcs}}
-            </span>
-          </a>
-        </div>
+        <app-teams-dropdown [inReport]="inReport" [settings]="settings"></app-teams-dropdown>
       </div>
       <div class="col-6 pl-1">
         <h5>Select Equipment</h5>
